requirements.txt
altair==6.0.0
anyio==4.12.0
attrs==25.4.0
av==16.0.1
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.14
ctranslate2==4.6.2
diskcache==5.6.3
exceptiongroup==1.3.1
faster-whisper==1.2.1
filelock==3.20.1
flatbuffers==25.12.19
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.3
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
llama_cpp_python==0.3.16
MarkupSafe==3.0.3
mpmath==1.3.0
narwhals==2.14.0
networkx==3.4.2
numpy==2.2.6
onnxruntime==1.23.2
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==6.33.2
pyarrow==22.0.0
pycparser==2.23
pydeck==0.9.1
pypiwin32==223
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pyttsx3==2.99
pytz==2025.2
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sounddevice==0.5.3
soundfile==0.13.1
streamlit==1.52.2
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.1
toml==0.10.2
torch==2.9.1
torchaudio==2.9.1
tornado==6.5.4
tqdm==4.67.1
typer-slim==0.21.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
watchdog==6.0.0
